Moved HasCase to its own module.
